Added tests for SubscriptionCancellationManager.get_earliest_possible_cancellation_date

+++ tapir/subscriptions/services/subscription_cancellation_manager.py
@@ -11,15 +11,10 @@ from tapir.wirgarten.utils import get_now
 class SubscriptionCancellationManager:
     @classmethod
     def get_earliest_possible_cancellation_date(cls, product: Product, member: Member):
-        if not get_parameter_value(Parameter.SUBSCRIPTION_AUTOMATIC_RENEWAL):
-            raise ImproperlyConfigured(
-                "Subscriptions don't have a cancellation date if automatic subscription renewal is disabled"
-            )
-
         subscriptions = (
             get_active_and_future_subscriptions()
             .filter(member=member, product=product)
-            .order_by("start_date")
+            .order_by("end_date")
         )
         for subscription in subscriptions:
             if TrialPeriodManager.is_subscription_in_trial(subscription):
@@ -33,9 +28,21 @@ class SubscriptionCancellationManager:
             product=product, member=member
         )
 
+        auto_renew_enabled = get_parameter_value(
+            Parameter.SUBSCRIPTION_AUTOMATIC_RENEWAL
+        )
+
         for subscription in get_active_and_future_subscriptions().filter(
             member=member, product=product
         ):
+            if (
+                not TrialPeriodManager.is_subscription_in_trial(subscription)
+                and not auto_renew_enabled
+            ):
+                raise ImproperlyConfigured(
+                    "Subscriptions outside of trial period can only be cancelled if auto renew is enabled"
+                )
+
             if subscription.start_date > cancellation_date:
                 subscription.delete()
                 continue
